Added multiple endpoints handled by single lambda
